requirement.txt
certifi @ file:///croot/certifi_1665076670883/work/certifi
PyOpenGL==3.1.6
